+++ src/data/navigationItems.ts
@@ -1,6 +1,6 @@
 export default [
-  { label: 'Home', path: '/' },
-  { label: 'Post a Job', path: '/post-job' },
+  { label: 'Home', path: '/offers' },
+  { label: 'Post a Job', path: '/offers/new' },
   { label: 'About Us', path: '/about' },
   { label: 'Contact', path: '/contact' },
 ];