+++ src/pages/offers/new/NewOffer.tsx
@@ -1,0 +1,5 @@
+function NewOffer() {
+  return <div>New Offer</div>;
+}
+
+export default NewOffer;