+++ src/pages/offers/new/index.ts
@@ -1,0 +1,3 @@
+import NewOffer from './NewOffer';
+
+export default NewOffer;